requirements.txt
groq
google-auth
google-auth-oauthlib
google-api-python-client
requests
python-dotenv